requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-hub
tavily-python
yfinance
wikipedia
streamlit
python-dotenv
groq
reportlab-4.4.10